#2135: final changes addressed
